Restrict Payment to only the necessary actions
It appears that the /v1/payments endpoints are temporary and will
eventually be combined into /v1/charges. In order to encourage writing
application code that will need as few updates as possible going
forward, we'll restrict what you can do via the Payment class to only
those actions that can't be performed via the Charge class.